TERA-432: Updating README. Moving RemoteStrategyFactory to its own file. Remote the DateUtil library and use the standard datetime library instead.
