+++ appcli/backup_manager/backup_manager.py
@@ -10,20 +10,15 @@ www.brightsparklabs.com
 """
 
 
+# standard libraries
 import os
 import re
 import shutil
 import tarfile
 from datetime import datetime, timezone
-
-# standard libraries
 from pathlib import Path
 
-# vendor libraries
-import dateutil.parser
-
 # local libraries
-from appcli.backup_manager.remote_strategy import AwsS3Strategy
 from appcli.logger import logger
 from appcli.models.cli_context import CliContext
 
@@ -293,7 +288,7 @@ class BackupManager:
         full_backup_directory = os.listdir(backup_dir)
         regex_pattern = re.compile(
             app_name.upper()
-            + "_\d{4}-[01]\d-[0-3]\dT[0-2]\d:[0-5]\d:[0-5]\d+([+-][0-2]\d:[0-5]\d|Z).tgz"
+            + "_\\d{4}-[01]\\d-[0-3]\\dT[0-2]\\d:[0-5]\\d:[0-5]\\d+([+-][0-2]\\d:[0-5]\\d|Z).tgz"
         )
         backup_files = [x for x in full_backup_directory if re.match(regex_pattern, x)]
 
@@ -304,7 +299,8 @@ class BackupManager:
             reverse=True,
         )
         backups_to_delete = backup_dir_files[
-            self.number_of_backups_to_retain - 1 :
+            self.number_of_backups_to_retain
+            - 1 :  # noqa: E203 - Disable flake8 error on spaces before a `:`
         ]  # -1 as we're 0 indexed
         for backup_to_delete in backups_to_delete:
             backup_file: Path = Path(os.path.join(backup_dir, backup_to_delete))
@@ -331,78 +327,8 @@ class BackupManager:
         sort_key = now.isoformat()
 
         try:
-            return dateutil.parser.parse(timestamp)
+            return datetime.fromisoformat(timestamp)
         except ValueError:
             logger.info(f"Unexpected backup found {filename}")
 
-        return dateutil.parser.parse(sort_key)
-
-
-class RemoteStrategyFactory:
-    """
-    Factory for getting all remote strategies that match the config
-    """
-
-    @staticmethod
-    def get_strategy(backup_manager, key_file: Path):
-        strategies = {"S3": AwsS3Strategy}
-
-        backup_strategies = []
-
-        for backup in backup_manager.remote_strategy_config:
-            cl = strategies.get(backup["type"], lambda: "Invalid remote strategy")
-
-            if isinstance(cl, str):
-                logger.error(
-                    f"No remote backup strategies found for type {backup['type']}"
-                )
-
-            strategy = cl(backup, key_file)
-
-            if RemoteStrategyFactory.__frequency_check(strategy.frequency):
-                backup_strategies.append(strategy)
-
-        return backup_strategies
-
-    def __frequency_check(frequency):
-        """
-        Check if today matches the pseudo cron `frequency`.
-        Frequency format is `* * *`:
-        First `*` is the day of the month.
-        Second `*` is the month.
-        Third `*` is the day of the week starting with monday=0.
-        `*` is a wildcard that is always matched.
-        e.g.
-        `* * *` Will always return True.
-        `* * 0` Will only return True on Mondays.
-
-        Args:
-            frequency: str. A frequency string that is checked to see if a backup should occur today.
-        Returns:
-            True if today matches the frequency, False if it does not.
-        """
-        if not len(frequency) == 5:
-            logger.error(f"Frequency string is invalid - {frequency}")
-            return False
-
-        toReturn = True
-
-        day_of_month = frequency[0]
-        month = frequency[2]
-        day_of_week = frequency[4]
-
-        today = datetime.today()
-
-        if day_of_month.isnumeric():
-            if not int(day_of_month) == today.day:
-                toReturn = False
-
-        if month.isnumeric():
-            if not int(month) == today.month:
-                toReturn = False
-
-        if day_of_week.isnumeric():
-            if not int(day_of_week) == today.weekday():
-                toReturn = False
-
-        return toReturn
+        return datetime.fromisoformat(sort_key)
